requirements.txt

pandas>=2.2.2
requests>=2.32.3
python-dotenv>=1.0.1
fastapi>=0.110.0
uvicorn>=0.29.0
pydantic>=2.7.0
google-cloud-bigquery>=3.25.0
reportlab>=4.2.0
